+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## v1.0.17 ~ 10-26-2020
 
-  * New Module: 1Pattern
+  * New Module: 1Pattern (basically one row of Patterns module)
+  * NoteSeqFu 'Start' now effects running sequence
   * Added 'Start' knobs and inputs to NoteSeq and NoteSeq16
   * Added 'Length' inputs to NoteSeq16